+++ src/extension.ts
@@ -9,6 +9,10 @@ import {
     demoteHeading,
     promoteHeading,
 } from "./headings";
+import {
+    promoteTree,
+    demoteTree,
+} from "./trees";
 
 export function activate(context: ExtensionContext) {
     const insertHeadingCmd = commands.registerTextEditorCommand("markdownOrgMode.insertHeading", insertHeading);
@@ -22,12 +26,18 @@ export function activate(context: ExtensionContext) {
     const demoteHeadingCmd = commands.registerTextEditorCommand("markdownOrgMode.demoteHeading", demoteHeading);
     const promoteHeadingCmd = commands.registerTextEditorCommand("markdownOrgMode.promoteHeading", promoteHeading);
 
+    const demoteTreeCmd = commands.registerTextEditorCommand("markdownOrgMode.demoteTree", demoteTree);
+    const promoteTreeCmd = commands.registerTextEditorCommand("markdownOrgMode.promoteTree", promoteTree);
+
     context.subscriptions.push(insertHeadingCmd);
     context.subscriptions.push(insertSubheadingCmd);
 
     context.subscriptions.push(toggleLineAndHeadingCmd);
     context.subscriptions.push(demoteHeadingCmd);
     context.subscriptions.push(promoteHeadingCmd);
+
+    context.subscriptions.push(demoteTreeCmd);
+    context.subscriptions.push(promoteTreeCmd);
 }
 
 export function deactivate() {}